+++ features/scripts/install_gradle_plugin.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 set -e
 
-./gradlew build install -x groovyDoc -x detekt -x test -PlocalVersion=9000.0.0-test
+./gradlew install -x detekt -x test -PlocalVersion=9000.0.0-test