+++ features/scripts/manual_upload_react_native.sh
@@ -1,0 +1,7 @@
+#!/usr/bin/env bash
+set -e
+
+cd $RN_FIXTURE_DIR
+echo "Test fixture used: $RN_FIXTURE_DIR, AGP=$AGP_VERSION, Gradle=$GRADLE_WRAPPER_VERSION"
+npm install
+./gradlew :app:clean :app:uploadBugsnagReleaseSourceMaps -x lint --stacktrace