+++ src/test/kotlin/com/bugsnag/android/gradle/PluginExtensionTest.kt
@@ -70,15 +70,18 @@ class PluginExtensionTest {
             assertTrue(uploadJvmMappings.get())
             assertNull(uploadNdkMappings.orNull)
             assertNull(uploadNdkUnityLibraryMappings.orNull)
+            assertNull(uploadReactNativeMappings.orNull)
             assertNull(sourceControl.repository.orNull)
             assertNull(sourceControl.revision.orNull)
             assertNull(sourceControl.provider.orNull)
+            assertNull(nodeModulesDir.orNull)
         }
 
         // ndk/unity upload defaults to false
         val plugin = proj.plugins.findPlugin(BugsnagPlugin::class.java)!!
         assertFalse(plugin.isUnityLibraryUploadEnabled(bugsnag, android))
         assertFalse(plugin.isNdkUploadEnabled(bugsnag, android))
+        assertFalse(plugin.isReactNativeUploadEnabled(bugsnag))
         assertEquals(emptyList<File>(), plugin.getSharedObjectSearchPaths(proj, bugsnag, android))
     }
 
@@ -104,9 +107,11 @@ class PluginExtensionTest {
             uploadJvmMappings.set(false)
             uploadNdkMappings.set(true)
             uploadNdkUnityLibraryMappings.set(true)
+            uploadReactNativeMappings.set(true)
             metadata.set(mapOf(Pair("test", "a")))
             objdumpPaths.set(mapOf(Pair("armeabi-v7a", "/test/foo")))
             sharedObjectPaths.set(listOf(File("/test/bar")))
+            nodeModulesDir.set(File("/test/foo/node_modules"))
 
             sourceControl.repository.set("https://github.com")
             sourceControl.revision.set("d0e98fc")
@@ -130,6 +135,7 @@ class PluginExtensionTest {
             assertEquals(mapOf(Pair("test", "a")), metadata.get())
             assertEquals(mapOf(Pair("armeabi-v7a", "/test/foo")), objdumpPaths.get())
             assertEquals(listOf(File("/test/bar")), sharedObjectPaths.get())
+            assertEquals(File("/test/foo/node_modules"), nodeModulesDir.get())
             assertEquals("https://github.com", sourceControl.repository.get())
             assertEquals("d0e98fc", sourceControl.revision.get())
             assertEquals("github", sourceControl.provider.get())
@@ -139,6 +145,8 @@ class PluginExtensionTest {
         val plugin = proj.plugins.findPlugin(BugsnagPlugin::class.java)!!
         assertTrue(plugin.isUnityLibraryUploadEnabled(bugsnag, android))
         assertTrue(plugin.isNdkUploadEnabled(bugsnag, android))
+        assertTrue(plugin.isReactNativeUploadEnabled(bugsnag))
+        assertEquals("http://localhost:1234", bugsnag.endpoint.get())
         val expected = listOf(
             File("/test/bar"),
             File(proj.projectDir, "src/main/jniLibs"),
@@ -167,6 +175,17 @@ class PluginExtensionTest {
             File(proj.rootDir, "unityLibrary/src/main/jniLibs")
         )
         assertEquals(expected, plugin.getSharedObjectSearchPaths(proj, bugsnag, android))
+    }
+
+    /**
+     * Verifies that in a project with the React property extension upload is disabled by default
+     */
+    @Test
+    fun reactNativeUploadHeuristics() {
+        val bugsnag = proj.extensions.getByType(BugsnagPluginExtension::class.java)
+        val plugin = proj.plugins.findPlugin(BugsnagPlugin::class.java)!!
+        proj.extensions.extraProperties.set("react", "some value")
+        assertFalse(plugin.isReactNativeUploadEnabled(bugsnag))
     }
 
     /**